sript_auto/requirement.txt
requests
os
dotenv 
csv
pandas
pathlib 
pymongo